Alter create word list service object to pass specs
Big thing here was to wrap the word creation in a transation, and fail
early. Then in errors, get each word error out of the word object,
rather than just have the default "words is invalid" message for
word_list.

This goes slightly away from the restful model, but we are in essence
abstracting away the fact that a wordlist has lots of words, into
wordlist just has an array of words, as that would be how the user would
edit the words.